Update 1:
Updated compute rules and assessment rules to have clean fluent unit tests. Implemented version change rules and added unit tests.
Version change rules has story in backlog that now should be code complete but not tested.
Update 2:
Updated convertToVersion to use string representation of version.

Related work items: #132635

+++ Bct.Workflow.Aggregates.CPPLib/include/BaseField.h
@@ -374,7 +374,7 @@ namespace Bct
                   }
                }
                std::string aggName = typeid(*_aggregate).name();
-               size_t size = _aggregate->MetaData().fieldInfo.size();
+               int32_t size = (int32_t)_aggregate->MetaData().fieldInfo.size();
                std::string fieldName = "unknown";
                std::string reqVersion = _aggregate->MetaData().versionInfo[_ver].Version();
                if (size > _fieldId)
